ref(issues): Rename filterThreadInfo to getThreadInfo
the function derives info from a thread, it doesn't filter anything.
While here, use the existing getFileName util instead of a local copy,
inline the getRelevantFrame helper (only used here), and type the
stacktrace properly so the any and the "this should not happen"
fallback go away.
