+++ static/app/components/events/interfaces/threads/threadSelector/index.tsx
@@ -13,7 +13,7 @@ import {trackAnalytics} from 'sentry/utils/analytics';
 import {unreachable} from 'sentry/utils/unreachable';
 import {useOrganization} from 'sentry/utils/useOrganization';
 
-import {filterThreadInfo, type ThreadInfo} from './filterThreadInfo';
+import {getThreadInfo, type ThreadInfo} from './getThreadInfo';
 import {Option} from './option';
 import {ThreadSelectorGrid, ThreadSelectorGridCell} from './styles';
 import {getMappedThreadState} from './threadStates';
@@ -37,7 +37,7 @@ const enum SortAttribute {
 }
 
 function getThreadLabel(
-  details: ReturnType<typeof filterThreadInfo>,
+  details: ReturnType<typeof getThreadInfo>,
   name: string | null | undefined
 ) {
   if (name?.length) {
@@ -64,7 +64,7 @@ export function ThreadSelector({
 
   const items = useMemo((): Array<SelectOptionOrSection<number>> => {
     const threadInfoMap = threads.reduce<Record<number, ThreadInfo>>((acc, thread) => {
-      acc[thread.id] = filterThreadInfo(event, thread, exception);
+      acc[thread.id] = getThreadInfo(event, thread, exception);
       return acc;
     }, {});
 
@@ -153,7 +153,7 @@ export function ThreadSelector({
             {t('Thread #%s: ', activeThread.id)}
             <ActiveThreadName>
               {getThreadLabel(
-                filterThreadInfo(event, activeThread, exception),
+                getThreadInfo(event, activeThread, exception),
                 activeThread.name
               )}
             </ActiveThreadName>
